Renamed bot directory
