+++ src/plugins/alanya-live-bot/server/bot/components/index.js
