+++ src/plugins/alanya-live-bot/server/bot/components/inlineCommands/enterCommand.js
