+++ src/plugins/alanya-live-bot/server/bot/components/inlineCommands/fullDescription.js
